html/semantics/embedded-content-0/media-elements/autoplay.html

<!doctype html>
<title>autoplaying fires play and playing events between canplay and canplaythrough</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<div id="log"></div>
<script>
function autoplay_test(tagName, src) {
  async_test(function(t) {
    var e = document.createElement(tagName);
    e.src = src;
    e.autoplay = true;
    document.body.appendChild(e);
    var actual_events = [];
    var expected_events = ['canplay', 'play', 'playing', 'canplaythrough'];
    expected_events.forEach(function(type) {
      e.addEventListener(type, t.step_func(function() {
        actual_events.push(type);
        if (type == 'canplaythrough') {
          assert_array_equals(actual_events, expected_events);
          t.done();
        }
      }));
    });
  }, document.title + ', ' + tagName);
}
autoplay_test('audio', getAudioURI('/media/sound_5'));
autoplay_test('video', getVideoURI('/media/movie_5'));
</script>
